fix(Content): size fill in cssSizeVariable, move component to /Content
